+++ backend/src/main/java/com/virtuallive/backend/model/entity/Comment.java
@@ -20,8 +20,8 @@ public class Comment {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "comment_id")
-    private Long commentId;
-    
+    private Integer commentId;
+
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "post_id")
     private Video post;
