+++ backend/src/main/java/com/virtuallive/backend/model/entity/Video.java
@@ -21,8 +21,8 @@ public class Video {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "post_id")
-    private Long postId;
-    
+    private Integer postId;
+
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "author_id", nullable = false)
     private User author;
