+++ backend/src/main/java/com/virtuallive/backend/repository/VideoRepository.java
@@ -10,8 +10,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface VideoRepository extends JpaRepository<Video, Long> {
-    
+public interface VideoRepository extends JpaRepository<Video, Integer> {
+
     Page<Video> findByIsDeletedFalseOrderByCreatedAtDesc(Pageable pageable);
     
     Page<Video> findByCategoryAndIsDeletedFalse(String category, Pageable pageable);
@@ -19,5 +19,5 @@ public interface VideoRepository extends JpaRepository<Video, Long> {
     @Query("SELECT v FROM Video v WHERE v.isDeleted = false ORDER BY v.views DESC")
     Page<Video> findPopularVideos(Pageable pageable);
     
-    List<Video> findByAuthor_UserIdAndIsDeletedFalse(Long authorId);
+    List<Video> findByAuthor_UserIdAndIsDeletedFalse(Integer authorId);
 }
